requirements.txt
firecrawl-py
python-dotenv
markdown2
xhtml2pdf
